Design and optimization of a pulsed power generator for electrical wire explosion
This master's thesis aims to develop robust techniques for inducing extreme conditions in fluids by employing a Pulsed Power Generator (PPG). The project focuses on studying and controlling electrically induced single-wire explosions to efficiently produce shock waves in air and water. An in-depth literature review is conducted, followed by an exploration of the necessary theoretical background to understand the multidisciplinary physical phenomena. A methodology is developed to ensure the safe operation of the PPG, along with defining the Schlieren optical setup and operating procedures for the high-speed camera capable of reaching speeds up to 10 million frames per second. The electrical parameters of the PPG are determined through short-circuit experiments and compared to an analogous RLC-like circuit using analytical and numerical simulations. Systematic experimentation is conducted across different copper wire diameters of 150 µm, 400 µm, and 500 µm and different initial capacitor voltages up to 23 kV. The optimal explosion conditions are identified, notably with a wire diameter of 400 µm in air, resulting in peak pressures in the order of hundreds of bar and Mach numbers up to 21.4. In water, the peak pressures reach tens of kilobar and Mach numbers up to 1.8. The analysis quantifies the transferred electrical energy and initial mechanical energy of the shock waves, reaching power magnitudes in the order of gigawatts and electrical-to-mechanical energy transfer efficiency up to 33.89%. The results are compared with numerical simulations and existing literature, culminating in a comprehensive report that synthesizes findings from the literature review, hands-on experimentation, and analysis.